CREATE PROC, ALTER, DROP PROC
Create the procedure spCountEmpMaritalStatus to count the number of employees by the department, the shift, the marial status and the gender. Three tables were connected using INNER JOIN to retrieve the necessary data.
Aggregation function COUNT () is used, GROUP BY clause if used .
sp_helptext was executed to see the text of the procedure.